requirements.txt
flask
flask-bootstrap
requests[security]
pyldap
